Moving the NetworkRequestTask in tasks.py and code refactoring
