+++ src/qgis_geonode/gui/geonode_map_layer_config_widget.py
@@ -21,6 +21,7 @@ from .. import (
     network,
     styles,
     utils,
+    tasks,
 )
 from ..apiclient import (
     base,
@@ -51,7 +52,7 @@ class GeonodeMapLayerConfigWidget(qgis.gui.QgsMapLayerConfigWidget, WidgetUi):
     upload_layer_pb: QtWidgets.QPushButton
     message_bar: qgis.gui.QgsMessageBar
 
-    network_task: typing.Optional[network.NetworkRequestTask]
+    network_task: typing.Optional[tasks.NetworkRequestTask]
     _apply_geonode_style: bool
     _apply_geonode_metadata: bool
     _layer_upload_api_client: typing.Optional[base.BaseGeonodeClient]
@@ -147,7 +148,7 @@ class GeonodeMapLayerConfigWidget(qgis.gui.QgsMapLayerConfigWidget, WidgetUi):
 
     def download_style(self):
         dataset = self.get_dataset()
-        self.network_task = network.NetworkRequestTask(
+        self.network_task = tasks.NetworkRequestTask(
             [network.RequestToPerform(QtCore.QUrl(dataset.default_style.sld_url))],
             self._api_client.network_requests_timeout,
             self.connection_settings.auth_config,
@@ -189,7 +190,7 @@ class GeonodeMapLayerConfigWidget(qgis.gui.QgsMapLayerConfigWidget, WidgetUi):
         if sld_data is not None:
             serialized_sld, content_type = sld_data
             dataset = self.get_dataset()
-            self.network_task = network.NetworkRequestTask(
+            self.network_task = tasks.NetworkRequestTask(
                 [
                     network.RequestToPerform(
                         QtCore.QUrl(dataset.default_style.sld_url),
@@ -322,7 +323,7 @@ class GeonodeMapLayerConfigWidget(qgis.gui.QgsMapLayerConfigWidget, WidgetUi):
     def upload_metadata(self) -> None:
         self.apply()
         current_metadata = self.layer.metadata()
-        self.network_task = network.NetworkRequestTask(
+        self.network_task = tasks.NetworkRequestTask(
             [
                 network.RequestToPerform(
                     QtCore.QUrl(self.get_dataset().link),
